Regular updates
Updated the squad status to only set squad as unavailable if a soldier is infiltrating (CI) and not when it is a normal CA
Added logic to avoid auto-filling of soldier slots in case of a CA deployment. Note that you can still select a squad and then add soldiers manually from barracks, where they will be labeled differently from those soldiers belonging to a different squad.
Added support for dynamic class icons using HL hook